Implement clipping function for all 5 planes

+++ SDL3D/Math.cpp
@@ -4,7 +4,7 @@
 #include <math.h>
 
 namespace Math {
-	std::vector<int> interpolateInt(float i0, float d0, float i1, float d1) {
+	std::vector<int> interpolateInt(int i0, float d0, int i1, float d1) {
 
 		if (i0 == i1) {
 			return std::vector<int> { static_cast<int>(d0) };
@@ -16,6 +16,7 @@ namespace Math {
 		std::vector<int> values{};
 
 		for (int x{ static_cast<int>(i0) }; x <= i1; ++x) {
+
 			values.emplace_back(static_cast<int>(c));
 			c += a;
 		}
