+++ SDL3D/Math.hpp
@@ -6,7 +6,7 @@
 namespace Math {
 	
 	std::vector<float> interpolateFloat(float i0, float d0, float i1, float d1);
-	std::vector<int> interpolateInt(float i0, float d0, float i1, float d1);
+	std::vector<int> interpolateInt(int i0, float d0, int i1, float d1);
 	float toRad(float const degrees);
